AI for enemy shooting

+++ Assets/Scripts/Data/Object/EnemyData.cs
@@ -3,5 +3,5 @@
 public class EnemyData
 {
     public const string name = "Enemy";
-    public float speed = 2f;
+    public int speed = 2;
 }
